Merge remote-tracking branch 'google_upstream/master' into develup-upstream-sync-191126

Conflicts:
	configure.py
	tensorflow/compiler/tests/BUILD
	tensorflow/compiler/tests/binary_ops_test.py
	tensorflow/compiler/tf2xla/BUILD
	tensorflow/compiler/xla/service/gpu/tests/BUILD
	tensorflow/compiler/xla/tests/BUILD
	tensorflow/compiler/xla/tests/build_defs.bzl
	tensorflow/core/kernels/fused_batch_norm_op.cc
	tensorflow/python/BUILD
	tensorflow/python/grappler/auto_mixed_precision_test.py
	tensorflow/python/keras/BUILD
	tensorflow/python/kernel_tests/BUILD
	tensorflow/python/kernel_tests/signal/BUILD
	tensorflow/python/ops/nn_ops.py
	tensorflow/python/ops/ragged/BUILD
	tensorflow/stream_executor/BUILD
	tensorflow/tensorflow.bzl
	tensorflow/tools/ci_build/builds/docker_test.sh
	tensorflow/tools/ci_build/ci_build.sh
	tensorflow/tools/ci_build/install/install_bazel.sh
	third_party/gpus/crosstool/hipcc_cc_toolchain_config.bzl.tpl
	third_party/gpus/rocm_configure.bzl

It looks like you may be committing a merge.
If this is not correct, please remove the file
	.git/MERGE_HEAD
and try again.

+++ tensorflow/lite/kernels/internal/reference/integer_ops/softmax.h
@@ -85,8 +85,9 @@ inline void Softmax(const SoftmaxParams& params,
             (shifted_scale * exp_in_0).raw(), num_bits_over_unit + 31 - 8);
         const int32 shifted_output = unsat_output - 128;
 
-        output_data[i * depth + c] =
-            static_cast<int8>(std::max(std::min(shifted_output, 127), -128));
+        output_data[i * depth + c] = static_cast<int8>(
+            std::max(std::min(shifted_output, static_cast<int32>(127)),
+                     static_cast<int32>(-128)));
 
       } else {
         output_data[i * depth + c] = -128;
